New features, editing/deleting aim, some bug fixed

+++ src/main/java/application/repositories/IAimRepository.java
@@ -1,7 +1,8 @@
 package application.repositories;
 
-import application.entities.Aim;
+import application.entities.aim.Aim;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IAimRepository extends CrudRepository<Aim, Integer> {
+public interface IAimRepository extends CrudRepository<Aim, Long> {
+
 }
